requirements.txt
astroid==3.3.8
dill==0.3.9
isort==5.13.2
mccabe==0.7.0
platformdirs==4.3.6
pylint==3.3.3
pyzmq==26.2.0
tomlkit==0.13.2
